refactor: move starter package
