+++ internal/tui/starter/model.go
